requirements.txt
streamlit
plotly
pillow
numpy